update Huggingface project according to new DPS template
