--- HuggingFace/Task1----environment-setup/README.md
@@ -1,17 +1,0 @@
-# Setting up the environment
-
-- Have Python installed on your machine. I am using python 3.9.
-
-- Download the files for the "**demo**" ai server which will be used in this tutorial by ->[clicking here](https://drive.google.com/uc?export=download&id=1xQ1KCpOrjRiJ51tLRMH6aBWi9tIBvS83)<-
-
-- cd into the downloaded directory
-
-```bash
-cd <Path>
-```
-
-- Install all the dependencies on your machine
-
-```bash
-pip3 install -r requirements.txt
-```
