--- HuggingFace/Task1----environment-setup/demo/models.py
@@ -1,4 +1,0 @@
-from pydantic import BaseModel
-
-class Text(BaseModel):
-    text: str
