--- HuggingFace/Task1----environment-setup/demo/save-t5.py
@@ -1,8 +1,0 @@
-from transformers import AutoTokenizer, AutoModelWithLMHead
-
-tokenizer = AutoTokenizer.from_pretrained("t5-small")
-
-model = AutoModelWithLMHead.from_pretrained("t5-small")
-
-model.save_pretrained('t5-small-saved')
-tokenizer.save_pretrained('t5-small-saved')
