--- HuggingFace/Task4----deploying-with-ci-cd/ai.yml
@@ -1,27 +1,0 @@
-name: ai Build
-
-on:
-  # run it on push to the staging repository branch
-  push:
-    branches:
-      - master
-    paths:
-      - "ai-fastapi-server/**"
-
-jobs:
-  build-frontend:
-    uses: ./.github/workflows/build-and-deploy.yml
-    with:
-      context: ./ai-fastapi-server
-      image-name: ai
-      deployment-name: ai-deployment
-      username: ${{ github.actor }}
-      environment: development
-      cluster-name: farmpedia
-      zone: europe-west2
-      project: batch16-projects
-      dockerfile: Dockerfile
-      repository: batch16--farmpedia
-    secrets:
-      GHCR_PAT: ${{ secrets.GHCR_PAT }}
-      REACT_SPRING_TEMPLATE_SA: ${{ secrets.REACT_SPRING_TEMPLATE_SA }}
