+++ HuggingFace/pack_envfile
@@ -1,0 +1,3 @@
+# This env file is related to building the AI container specifying the python version 3.8 and setting an entrypoint for the container
+GOOGLE_RUNTIME_VERSION=3.8 
+GOOGLE_ENTRYPOINT=python ./inference.py
